改善: サーバのapp構成をモジュール分割
